Created new file: index.js, folder - countries
